organized train dir
